First step toward fixing the "unbound location" bug (18h, 25f, 14a, 14b work)
In InferLocations, tryInRegion and tryNeedRegion process the set of
constraints and decide if they need to add a StartRegionL constraint
to bind a location to the start a fresh region. While doing
so, they ensure that there isn't another StartRegionL constraint
(noRegionStart) which already binds that same location.

Instead of merely checking the presence of *some* StartRegionL
constraint, noRegionStart should check if that specific location is bound --
otherwise the parent fn can allocate a new region. It must also
process *all* the constraints passed in, to ensure that there's
exactly one StartRegionL constraint per location.